Fix "launchOrFocus" behavior for Atom
Some time in the last few weeks, something seems to have changed such
that Hyper+E always opens a *new* Atom instance, even when there's
already an exising Atom instance running. I would expect
hs.application.launchOrFocus to focus the currently-running Atom
instance, but that's not happening for some reason. This change seems to
be a successful workaround for this issue.