Rename option* -> fn* to reflect key name
I saw that the contents of the optionHandler function were updated to
watch for presses of the "fn" key instead of the "alt" key. At that time
the function (and its corresponding functions) were not renamed. This
commit aims to make the name of the function reflect its purpose. I
acknowledge that `fnHandler` is potentially misleading as `fn` is a
common abbreviation for function. I'm open to this name changing.